move object types and compiler into schema module
