+++ src/dsl/mod.rs
@@ -1,3 +1,1 @@
-pub mod compiler;
-pub mod object_types;
 pub mod schema;
