+++ src/dsl/schema/deserialization.rs
@@ -2,10 +2,10 @@ use serde::de::Error;
 use serde_yaml::Mapping;
 use serde_yaml::Value;
 
-use crate::dsl::compiler::CompilationError;
-use crate::dsl::object_types::deserialization::deserialize_object_type;
-use crate::dsl::object_types::ObjectType;
-use crate::dsl::object_types::RawObjectType;
+use crate::dsl::schema::compiler::CompilationError;
+use crate::dsl::schema::object_types::deserialization::deserialize_object_type;
+use crate::dsl::schema::object_types::ObjectType;
+use crate::dsl::schema::object_types::RawObjectType;
 use crate::dsl::schema::Property;
 use crate::dsl::schema::PropertyEntry;
 use crate::dsl::schema::PropertyList;
