+++ src/dsl/schema/object_types/deserialization.rs
@@ -2,12 +2,12 @@ use serde::de::Error;
 use serde_yaml::Mapping;
 use serde_yaml::Value;
 
-use crate::dsl::object_types::bounds::deserialization::deserialize_array_object_bounds;
-use crate::dsl::object_types::bounds::deserialization::deserialize_boolean_object_bounds;
-use crate::dsl::object_types::bounds::deserialization::deserialize_integer_bounds;
-use crate::dsl::object_types::bounds::deserialization::deserialize_string_object_bounds;
-use crate::dsl::object_types::ObjectType;
-use crate::dsl::object_types::RawObjectType;
+use crate::dsl::schema::object_types::bounds::deserialization::deserialize_array_object_bounds;
+use crate::dsl::schema::object_types::bounds::deserialization::deserialize_boolean_object_bounds;
+use crate::dsl::schema::object_types::bounds::deserialization::deserialize_integer_bounds;
+use crate::dsl::schema::object_types::bounds::deserialization::deserialize_string_object_bounds;
+use crate::dsl::schema::object_types::ObjectType;
+use crate::dsl::schema::object_types::RawObjectType;
 
 pub fn deserialize_object_type<E>(mapping: &Mapping) -> Result<Option<Vec<ObjectType>>, E>
 where
