+++ src/output/generator.rs
@@ -1,8 +1,8 @@
 use serde_json::Map;
 
-use crate::dsl::compiler::CompilationError;
-use crate::dsl::compiler::compile;
-use crate::dsl::compiler::CompiledSchema;
+use crate::dsl::schema::compiler::CompilationError;
+use crate::dsl::schema::compiler::compile;
+use crate::dsl::schema::compiler::CompiledSchema;
 use crate::output::JsonSchema;
 use crate::output::UiObject;
