+++ src/output/serialization/object_types.rs
@@ -6,16 +6,16 @@ use serde::ser::SerializeMap;
 use serde::Serialize;
 use serde::Serializer;
 
-use crate::dsl::object_types::bounds::ArrayItemObjectBounds;
-use crate::dsl::object_types::bounds::ArrayObjectBounds;
-use crate::dsl::object_types::bounds::ArrayUniqueItemBound;
-use crate::dsl::object_types::bounds::BooleanObjectBounds;
-use crate::dsl::object_types::bounds::EnumerationValue;
-use crate::dsl::object_types::bounds::IntegerBound;
-use crate::dsl::object_types::bounds::IntegerObjectBounds;
-use crate::dsl::object_types::bounds::StringLength;
-use crate::dsl::object_types::bounds::StringObjectBounds;
-use crate::dsl::object_types::RawObjectType;
+use crate::dsl::schema::object_types::bounds::ArrayItemObjectBounds;
+use crate::dsl::schema::object_types::bounds::ArrayObjectBounds;
+use crate::dsl::schema::object_types::bounds::ArrayUniqueItemBound;
+use crate::dsl::schema::object_types::bounds::BooleanObjectBounds;
+use crate::dsl::schema::object_types::bounds::EnumerationValue;
+use crate::dsl::schema::object_types::bounds::IntegerBound;
+use crate::dsl::schema::object_types::bounds::IntegerObjectBounds;
+use crate::dsl::schema::object_types::bounds::StringLength;
+use crate::dsl::schema::object_types::bounds::StringObjectBounds;
+use crate::dsl::schema::object_types::RawObjectType;
 
 impl Serialize for EnumerationValue {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
